requirements.txt
anyio>=4.6.2.post1
Brotli>=1.1.0
certifi>=2024.8.30
charset-normalizer>=3.4.0
h11>=0.14.0
httpcore>=1.0.6
httpx>=0.27.2
idna>=3.10
mutagen>=1.47.0
ollama>=0.3.3
pip>=24.2
pycryptodomex>=3.21.0
requests>=2.32.3
sniffio>=1.3.1
urllib3>=2.2.3
websockets>=13.1
webvtt-py>=0.5.1
yt-dlp>=2024.10.7
pycairo>=1.27.0
pygobject>=3.50.0
